+++ subprojects/calendar/src/main/java/nl/tudelft/sem/calendar/repositories/LectureRepository.java
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LectureRepository extends JpaRepository<Lecture, Long> {
+public interface LectureRepository extends JpaRepository<Lecture, Integer> {
 
     List<Lecture> findAll();
-    List<Lecture> findByLectureId(long lectureId);
+    List<Lecture> findByLectureId(int lectureId);
     List<Lecture> findByCourseId(String courseId);
     List<Lecture> findByDate(LocalDate date);
 }
